bug: split stream orchestration and relax test file-length gate
Changes Summary:
- extract stream loop and event handlers into agent_streaming.py
- store per-stream baseline state for interruption cleanup paths
- add regression coverage for callback failures, retry cleanup, and failing stream logs
- exempt tests/ from the non-test Python file-length pre-commit rule and update docs
- record execution notes under .artifacts/execute

Validation:
- uv run pytest tests/unit/core/test_request_orchestrator_parallel_tools.py tests/unit/core/test_stream_phase_debug.py
