+++ .artifacts/execute/2026-04-08_18-35-53_stream-non-abort-exception-cleanup.md
@@ -1,0 +1,87 @@
+---
+title: "stream non-abort exception cleanup execution log"
+link: "stream-non-abort-exception-cleanup-execute"
+type: debug_history
+ontological_relations:
+  - relates_to: [[stream-non-abort-exception-cleanup-plan]]
+tags: [execute, stream-exceptions]
+uuid: "1f44af67-934a-4c54-9cd4-352d436f2eb6"
+created_at: "2026-04-08T18:35:53-05:00"
+owner: "fabian"
+plan_path: ".artifacts/plan/2026-04-08_18-31-25_stream-non-abort-exception-cleanup/PLAN.md"
+start_commit: "91a9022a"
+end_commit: ""
+env: {target: "local", notes: "Stopped after repo-level gate failures outside plan scope."}
+---
+
+## Pre-Flight Checks
+- Branch: `master`
+- Rollback: `477402c1`
+- DoR: satisfied
+- Ready: yes
+- Access/secrets: present
+- Fixtures/data: ready
+
+## Task Execution
+
+### T001 - bug: store per-stream baseline and centralize interruption cleanup
+- Status: completed
+- Commit: not created
+- Files: `src/tunacode/core/agents/helpers.py`, `src/tunacode/core/agents/main.py`, `tests/unit/core/test_request_orchestrator_parallel_tools.py`
+- Commands: `uv run pytest tests/unit/core/test_request_orchestrator_parallel_tools.py -k 'uses_active_stream_baseline_for_cleanup or reraises_non_abort_exception_after_stream_cleanup or callback_failure_cleanup or retry_stream_cleanup_baseline'` -> `5 passed, 9 deselected in 0.68s`
+- Tests: pass
+- Coverage delta: not measured
+- Notes: Added `baseline_message_count` to `_TinyAgentStreamState` and centralized interrupted-stream cleanup behind a shared helper.
+
+### T002 - bug: route unexpected stream exceptions through cleanup for every stream attempt
+- Status: completed
+- Commit: not created
+- Files: `src/tunacode/core/agents/main.py`, `tests/unit/core/test_request_orchestrator_parallel_tools.py`
+- Commands: `uv run pytest tests/unit/core/test_request_orchestrator_parallel_tools.py -k 'uses_active_stream_baseline_for_cleanup or reraises_non_abort_exception_after_stream_cleanup or callback_failure_cleanup or retry_stream_cleanup_baseline'` -> `5 passed, 9 deselected in 0.68s`
+- Tests: pass
+- Coverage delta: not measured
+- Notes: Added `_run_stream_with_cleanup()` so unexpected in-stream exceptions clean up only while active stream state is live and then re-raise the original exception.
+
+### T003 - test: add regression coverage for tool callback failures during stream tool events
+- Status: completed
+- Commit: not created
+- Files: `tests/unit/core/test_request_orchestrator_parallel_tools.py`
+- Commands: `uv run pytest tests/unit/core/test_request_orchestrator_parallel_tools.py -k 'uses_active_stream_baseline_for_cleanup or reraises_non_abort_exception_after_stream_cleanup or callback_failure_cleanup or retry_stream_cleanup_baseline'` -> `5 passed, 9 deselected in 0.68s`
+- Tests: pass
+- Coverage delta: not measured
+- Notes: Added callback-failure regressions for both `tool_start_callback` and `tool_result_callback`, including interrupted partial output and preserved completed results.
+
+### T004 - test: add retry-path regression coverage for cleanup baseline correctness
+- Status: completed
+- Commit: not created
+- Files: `tests/unit/core/test_request_orchestrator_parallel_tools.py`
+- Commands: `uv run pytest tests/unit/core/test_request_orchestrator_parallel_tools.py -k 'uses_active_stream_baseline_for_cleanup or reraises_non_abort_exception_after_stream_cleanup or callback_failure_cleanup or retry_stream_cleanup_baseline'` -> `5 passed, 9 deselected in 0.68s`
+- Tests: pass
+- Coverage delta: not measured
+- Notes: Added retry-stream failure coverage proving cleanup uses the retry stream baseline rather than the original request baseline.
+
+### T005 - test: add failing-stream logging regression coverage for message-update exceptions
+- Status: completed
+- Commit: not created
+- Files: `tests/unit/core/test_stream_phase_debug.py`
+- Commands: `uv run pytest tests/unit/core/test_stream_phase_debug.py -k failing_stream_logging` -> `1 passed, 1 deselected in 0.66s`
+- Tests: pass
+- Coverage delta: not measured
+- Notes: Added failing-stream logging coverage asserting `_run_stream()` omits success end lines when a streaming callback raises.
+
+## Gate Results
+- Tests: pass -> `uv run pytest` => `325 passed, 2 skipped in 6.30s`
+- Coverage: fail -> `uv run coverage report` => `No source for code: '/home/fabian/tunacode/src/tunacode/configuration/feature_flags.py'`
+- Type checks: pass -> `uv run mypy src/` => `Success: no issues found in 172 source files`
+- Security: not run
+- Linters: fail -> `uv run black --check src/` => `87 files would be reformatted, 85 files would be left unchanged`
+
+## Issues & Resolutions
+- Gate C - `black --check src/` failed on many pre-existing files outside this plan slice. No remediation attempted because bulk repo formatting is out of scope.
+- Gate C - `coverage report` failed because coverage data references missing file `src/tunacode/configuration/feature_flags.py`. No remediation attempted because it is unrelated to the stream-cleanup plan and the execution workflow requires stopping for user direction after gate failures.
+
+## Success Criteria
+- [ ] All planned gates passed
+- [ ] Rollout completed or rolled back
+- [x] KPIs/SLOs within thresholds
+- [x] Execution log saved
